# Back/routes/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"TypeContrat": {
				"enum": [
					"CDI",
					"CDD",
					"stage"
				],
				"type": "string"
			},
			"Role": {
				"enum": [
					"Admin",
					"RH",
					"Superviseur",
					"Employe"
				],
				"type": "string"
			},
			"UserOutput": {
				"properties": {
					"id_user": {
						"type": "number",
						"format": "double"
					},
					"matricule": {
						"type": "string"
					},
					"nom": {
						"type": "string"
					},
					"prenom": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"badge": {
						"type": "string"
					},
					"empreinte": {
						"type": "string"
					},
					"poste": {
						"type": "string"
					},
					"type_contrat": {
						"$ref": "#/components/schemas/TypeContrat"
					},
					"date_embauche": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"date_fin_contrat": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"id_lieu": {
						"type": "number",
						"format": "double"
					},
					"id_equipe": {
						"type": "number",
						"format": "double"
					},
					"id_departement": {
						"type": "number",
						"format": "double"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					}
				},
				"required": [
					"id_user",
					"matricule",
					"nom",
					"email",
					"phone",
					"badge",
					"poste",
					"type_contrat",
					"date_embauche",
					"id_lieu",
					"id_equipe",
					"id_departement",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"TypePointage": {
				"enum": [
					"entree",
					"sortie"
				],
				"type": "string"
			},
			"ModePointage": {
				"enum": [
					"bio",
					"badge",
					"manuel"
				],
				"type": "string"
			},
			"StatutPointage": {
				"enum": [
					"normal",
					"retard",
					"avance"
				],
				"type": "string"
			},
			"PointageOutput": {
				"properties": {
					"id_pointage": {
						"type": "number",
						"format": "double"
					},
					"type": {
						"$ref": "#/components/schemas/TypePointage"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"mode": {
						"$ref": "#/components/schemas/ModePointage"
					},
					"statut": {
						"$ref": "#/components/schemas/StatutPointage"
					},
					"id_pointeuse": {
						"type": "number",
						"format": "double"
					},
					"matricule": {
						"type": "string"
					},
					"serialNo": {
						"type": "number",
						"format": "double"
					},
					"pointeuse": {
						"properties": {
							"lieu": {},
							"id_lieu": {
								"type": "number",
								"format": "double"
							},
							"pointeuse": {
								"type": "string"
							},
							"adresse_ip": {
								"type": "string"
							},
							"id_pointeuse": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"lieu",
							"id_lieu",
							"adresse_ip",
							"id_pointeuse"
						],
						"type": "object"
					},
					"user": {
						"properties": {
							"departement": {},
							"equipe": {},
							"lieu": {},
							"role": {
								"type": "string"
							},
							"id_departement": {
								"type": "number",
								"format": "double"
							},
							"id_equipe": {
								"type": "number",
								"format": "double"
							},
							"id_lieu": {
								"type": "number",
								"format": "double"
							},
							"date_fin_contrat": {
								"anyOf": [
									{
										"type": "string",
										"format": "date-time"
									},
									{
										"type": "string"
									}
								]
							},
							"date_embauche": {
								"anyOf": [
									{
										"type": "string",
										"format": "date-time"
									},
									{
										"type": "string"
									}
								]
							},
							"type_contrat": {
								"type": "string"
							},
							"poste": {
								"type": "string"
							},
							"empreinte": {
								"type": "string"
							},
							"badge": {
								"type": "string"
							},
							"phone": {
								"type": "string"
							},
							"email": {
								"type": "string"
							},
							"prenom": {
								"type": "string"
							},
							"nom": {
								"type": "string"
							},
							"matricule": {
								"type": "string"
							},
							"id_user": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"role",
							"id_departement",
							"id_equipe",
							"id_lieu",
							"date_embauche",
							"type_contrat",
							"poste",
							"badge",
							"phone",
							"email",
							"nom",
							"matricule",
							"id_user"
						],
						"type": "object"
					}
				},
				"required": [
					"id_pointage",
					"type",
					"date",
					"mode",
					"statut",
					"id_pointeuse",
					"matricule",
					"serialNo"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PointageValidationResult": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IPointage.type-or-date-or-mode-or-statut-or-id_pointeuse-or-matricule-or-serialNo_": {
				"properties": {
					"matricule": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/TypePointage"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"mode": {
						"$ref": "#/components/schemas/ModePointage"
					},
					"statut": {
						"$ref": "#/components/schemas/StatutPointage"
					},
					"id_pointeuse": {
						"type": "number",
						"format": "double"
					},
					"serialNo": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"matricule",
					"type",
					"date",
					"mode",
					"statut",
					"id_pointeuse",
					"serialNo"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"PointageCreationParams": {
				"$ref": "#/components/schemas/Pick_IPointage.type-or-date-or-mode-or-statut-or-id_pointeuse-or-matricule-or-serialNo_"
			},
			"Partial_PointageCreationParams_": {
				"properties": {
					"matricule": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/TypePointage"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"mode": {
						"$ref": "#/components/schemas/ModePointage"
					},
					"statut": {
						"$ref": "#/components/schemas/StatutPointage"
					},
					"id_pointeuse": {
						"type": "number",
						"format": "double"
					},
					"serialNo": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"PointageUpdateParams": {
				"$ref": "#/components/schemas/Partial_PointageCreationParams_"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "back-rh-connect",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/users/matricule/{matricule}": {
			"get": {
				"operationId": "GetUserByMatricule",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserOutput"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "matricule",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/pointages/{id}": {
			"get": {
				"operationId": "GetPointage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PointageOutput"
								}
							}
						}
					}
				},
				"tags": [
					"Pointage"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdatePointage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PointageValidationResult"
								}
							}
						}
					}
				},
				"tags": [
					"Pointage"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PointageUpdateParams"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeletePointage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PointageOutput"
								}
							}
						}
					}
				},
				"tags": [
					"Pointage"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/pointages": {
			"post": {
				"operationId": "CreatePointage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PointageValidationResult"
								}
							}
						}
					}
				},
				"tags": [
					"Pointage"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PointageCreationParams"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllPointages",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PointageOutput"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Pointage"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH",
							"superviseur"
						]
					}
				],
				"parameters": []
			}
		},
		"/pointages/matricule/{matricule}": {
			"get": {
				"operationId": "GetPointagesByMatricule",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PointageOutput"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Pointage"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "matricule",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/pointages/date/{date}": {
			"get": {
				"operationId": "GetPointagesByDate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PointageOutput"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Pointage"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH",
							"superviseur"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "date",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/pointages/matricule/{matricule}/date/{date}": {
			"get": {
				"operationId": "GetPointagesByMatriculeAndDate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PointageOutput"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Pointage"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "matricule",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "date",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/pointages/pointeuse/{id_pointeuse}": {
			"get": {
				"operationId": "GetPointagesByPointeuse",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PointageOutput"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Pointage"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH",
							"superviseur"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id_pointeuse",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/pointages/import-pointages-api": {
			"post": {
				"operationId": "ImportPointagesFromApi",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"errors": {
											"items": {
												"type": "string"
											},
											"type": "array"
										},
										"totalPointages": {
											"type": "number",
											"format": "double"
										},
										"count": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"count",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Pointage"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH"
						]
					}
				],
				"parameters": []
			}
		},
		"/pointages/import-pointages-api-date": {
			"post": {
				"operationId": "ImportPointagesFromApiByDate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"errors": {
											"items": {
												"type": "string"
											},
											"type": "array"
										},
										"totalPointages": {
											"type": "number",
											"format": "double"
										},
										"count": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"count",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Pointage"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"date": {
										"type": "string"
									}
								},
								"required": [
									"date"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}